+++ subscription/util/blocking/catchup-subscription/src/test/java/org/occurrent/subscription/blocking/durable/catchup/CatchupSubscriptionModelTest.java
@@ -33,6 +33,7 @@ import org.junit.jupiter.api.extension.RegisterExtension;
 import org.occurrent.domain.DomainEvent;
 import org.occurrent.domain.NameDefined;
 import org.occurrent.domain.NameWasChanged;
+import org.occurrent.eventstore.api.SortBy;
 import org.occurrent.eventstore.mongodb.nativedriver.EventStoreConfig;
 import org.occurrent.eventstore.mongodb.nativedriver.MongoEventStore;
 import org.occurrent.mongodb.timerepresentation.TimeRepresentation;
@@ -63,6 +64,7 @@ import static java.util.Objects.requireNonNull;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
 import static org.awaitility.Durations.FIVE_SECONDS;
+import static org.occurrent.filter.Filter.TIME;
 import static org.occurrent.filter.Filter.type;
 import static org.occurrent.functional.CheckedFunction.unchecked;
 import static org.occurrent.subscription.OccurrentSubscriptionFilter.filter;
@@ -153,6 +155,34 @@ public class CatchupSubscriptionModelTest {
             assertThat(state).hasSize(2);
             assertThat(state).extracting(CloudEvent::getType).containsOnly(NameDefined.class.getName());
         });
+    }
+
+    @Test
+    void catchup_subscription_reads_historic_events_with_filter_using_custom_sort_by_during_catchup_phase() {
+        // Given
+        LocalDateTime now = LocalDateTime.now();
+        String eventId1 = UUID.randomUUID().toString();
+        String eventId2 = UUID.randomUUID().toString();
+        String eventId3 = UUID.randomUUID().toString();
+        NameDefined nameDefined1 = new NameDefined(eventId1, now, "name1");
+        NameDefined nameDefined2 = new NameDefined(eventId2, now.plusSeconds(2), "name2");
+        NameWasChanged nameWasChanged1 = new NameWasChanged(eventId3, now.plusSeconds(10), "name3");
+
+        mongoEventStore.write("1", 0, serialize(nameDefined1));
+        mongoEventStore.write("2", 0, serialize(nameDefined2));
+        mongoEventStore.write("1", 1, serialize(nameWasChanged1));
+
+        CopyOnWriteArrayList<CloudEvent> state = new CopyOnWriteArrayList<>();
+
+        CatchupSubscriptionModelConfig cfg = new CatchupSubscriptionModelConfig(100, useSubscriptionPositionStorage(storage).andPersistSubscriptionPositionDuringCatchupPhaseForEveryNEvents(1))
+                .catchupPhaseSortBy(SortBy.descending(TIME));
+        subscription = newCatchupSubscription(database, eventCollection, TimeRepresentation.DATE, cfg);
+
+        // When
+        subscription.subscribe(UUID.randomUUID().toString(), StartAt.subscriptionPosition(TimeBasedSubscriptionPosition.beginningOfTime()), state::add).waitUntilStarted();
+
+        // Then
+        await().atMost(FIVE_SECONDS).with().pollInterval(Duration.of(20, MILLIS)).untilAsserted(() -> assertThat(state).extracting(CloudEvent::getId).containsExactly(eventId3, eventId2, eventId1));
     }
 
     @Test